backend/requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Gmail API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Google Cloud Storage
google-cloud-storage==2.14.0

# Resume Processing
PyPDF2==3.0.1
PyMuPDF==1.23.8
python-docx==1.1.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0

# AI Integration
openai==1.6.1
anthropic==0.8.1
requests==2.31.0

# Utilities
python-magic==0.4.27